Clean up LocationDrawings component
